requirements.txt
python>=3.10

# CLI 界面
rich>=13.0.0
prompt-toolkit>=3.0.0

# WinDBG 集成
# pykd>=0.3.0  # 可选，如果 pykd 不可用则使用 subprocess

# LLM 集成
openai>=1.0.0
# anthropic>=0.5.0  # 可选

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 异步支持
aiohttp>=3.8.0

# 工具库
click>=8.0.0
colorama>=0.4.0

# 开发依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
